+++ api/src/main/java/com/hcs/config/advisor/ExceptionAdvisor.java
@@ -4,6 +4,7 @@ import com.hcs.config.advisor.result.ExceptionResult;
 import com.hcs.config.advisor.result.ValidationResult;
 import com.hcs.dto.response.HcsResponse;
 import com.hcs.dto.response.HcsResponseManager;
+import com.hcs.dto.response.method.HcsException;
 import com.hcs.exception.ErrorCode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
@@ -32,6 +33,9 @@ public class ExceptionAdvisor {
     @Autowired
     private HcsResponseManager hcsResponseManager;
 
+    @Autowired
+    private HcsException hcsException;
+
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ValidationResult handleBindException(BindException bindException, Locale locale) {
@@ -42,7 +46,8 @@ public class ExceptionAdvisor {
     @ExceptionHandler(NumberFormatException.class)
     public HcsResponse NumberFormatExceptionHandler() {
         ErrorCode error = ErrorCode.NUMBER_FORMAT;
-        return hcsResponseManager.Exception(error.getStatus(), new ExceptionResult(error.getErrorCode(), error.getMessage()));
+//        return hcsResponseManager.Exception(error.getStatus(), new ExceptionResult(error.getErrorCode(), error.getMessage()));
+        return hcsResponseManager.makeHcsResponse(hcsException.exception(error.getStatus(), new ExceptionResult(error.getErrorCode(), error.getMessage())));
     }
 
     // TODO 추후 Response가 만들어지면 공통으로 처리될 error에 대한 전역적인 @ExceptionHandler 추가 작성될 것임.
