+++ api/src/main/java/com/hcs/controller/ChatMessageController.java
@@ -2,9 +2,10 @@ package com.hcs.controller;
 
 import com.hcs.domain.ChatMessage;
 import com.hcs.domain.User;
-import com.hcs.dto.ChatMessageDto;
+import com.hcs.dto.request.ChatMessageDto;
 import com.hcs.dto.response.HcsResponse;
 import com.hcs.dto.response.HcsResponseManager;
+import com.hcs.dto.response.method.HcsSubmit;
 import com.hcs.service.ChatMessageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.MessageMapping;
@@ -22,6 +23,7 @@ public class ChatMessageController {
     private final SimpMessagingTemplate template;
     private final HcsResponseManager hcsResponseManager;
     private final ChatMessageService chatMessageService;
+    private final HcsSubmit submit;
 
     @MessageMapping("/chat/message") // pub
     public HcsResponse message(ChatMessageDto message) {
@@ -35,6 +37,6 @@ public class ChatMessageController {
 
         ChatMessage newChat = chatMessageService.createChatMessage(message);
 
-        return hcsResponseManager.submit.chatMessage(newChat.getId());
+        return hcsResponseManager.makeHcsResponse(submit.chatMessage(newChat.getId()));
     }
 }
