+++ api/src/main/java/com/hcs/controller/ClubController.java
@@ -1,9 +1,11 @@
 package com.hcs.controller;
 
 import com.hcs.domain.Club;
-import com.hcs.dto.ClubDto;
+import com.hcs.dto.request.ClubDto;
 import com.hcs.dto.response.HcsResponse;
 import com.hcs.dto.response.HcsResponseManager;
+import com.hcs.dto.response.method.HcsInfo;
+import com.hcs.dto.response.method.HcsSubmit;
 import com.hcs.service.ClubService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,25 +29,28 @@ public class ClubController {
 
     private final ClubService clubService;
     private final HcsResponseManager responseManager;
+    private final HcsInfo info;
+    private final HcsSubmit submit;
 
     @PostMapping("/submit")
     public HcsResponse createClub(@Valid @RequestBody ClubDto clubDto, HttpServletRequest request) {
         //TODO : 로그인한 유저인지 검증 추가
 
         Club newClub = clubService.saveNewClub(clubDto);
-        return responseManager.submit.club(newClub.getId(), getBaseUrl(request));
+        return responseManager.makeHcsResponse(submit.club(newClub.getId(), getBaseUrl(request)));
     }
 
     @GetMapping("/info")
     public HcsResponse clubInfo(@RequestParam("clubId") Long id, HttpServletRequest request) {
         Club club = clubService.getClub(id);
-        return responseManager.info.club(club, getBaseUrl(request));
+        return responseManager.makeHcsResponse(info.club(club, getBaseUrl(request)));
+
     }
 
     private String getBaseUrl(HttpServletRequest request) {
         return request.getRequestURL().toString().replace(request.getRequestURI(), "") + "/";
     }
-    
+
     //TODO : club list
 //    @GetMapping("/list")
 //    public HcsResponse clubList(@RequestParam("page")int page,@RequestParam("category") String category) {
