+++ api/src/main/java/com/hcs/controller/CommentController.java
@@ -2,8 +2,8 @@ package com.hcs.controller;
 
 import com.hcs.domain.TradePost;
 import com.hcs.domain.User;
-import com.hcs.dto.CommentDto;
 import com.hcs.dto.RespCommentDto;
+import com.hcs.dto.request.CommentDto;
 import com.hcs.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
