--- api/src/main/java/com/hcs/exception/HcsException.java
@@ -1,4 +1,0 @@
-package com.hcs.exception;
-
-public interface HcsException {
-}
