+++ api/src/main/java/com/hcs/exception/HcsExceptionType.java
@@ -1,0 +1,4 @@
+package com.hcs.exception;
+
+public interface HcsExceptionType {
+}
