+++ api/src/test/java/com/hcs/controller/CommentControllerTest.java
@@ -2,7 +2,7 @@ package com.hcs.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hcs.config.EnableMockMvc;
-import com.hcs.dto.CommentDto;
+import com.hcs.dto.request.CommentDto;
 import com.jayway.jsonpath.JsonPath;
 import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.junit.jupiter.api.DisplayName;
