+++ api/src/test/java/com/hcs/service/ClubServiceTest.java
@@ -1,7 +1,7 @@
 package com.hcs.service;
 
 import com.hcs.domain.Club;
-import com.hcs.dto.ClubDto;
+import com.hcs.dto.request.ClubDto;
 import com.hcs.mapper.ClubMapper;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
