+++ api/src/test/java/com/hcs/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package com.hcs.service;
 
 import com.hcs.domain.User;
-import com.hcs.dto.SignUpDto;
+import com.hcs.dto.request.SignUpDto;
 import com.hcs.mapper.UserMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
